Improve the MTU test.
This patch verifies that the dnsmasq_config_file was properly set,
since the CloudbaseInit plugins (MTU and NTP) are relying on information
retrieved from DHCP in order to do their job. In order to do so, one
must add the option 'dnsmasq_config_file=<file>' to
neutron/dhcp_agent.ini and configure inside that file various DHCP
options, through dhcp-option-force. The patch improves the lookup
in the dnsmasq_config_file for the MTU value.