Reset the excepthook changed by tempest.
tempest sets its own excepthook, which will log the error
using the tempest logger. Unfortunately, we are not using
the tempest logger, so any uncaught error goes into nothingness.
The code which sets the excepthook is here:
https://github.com/openstack/tempest/blob/master/tempest/openstack/common/log.py#L420
Since the isolated credentials provider imports tempest.config, we
import it already here, so that it sets the excepthook it wants
and then we reset the value to the original excepthook.