Add some support for image matrix testing.
First, this patch drops the use of oslo.config, using ConfigParser instead,
in order to have support for image specific options in the config file,
which is currently not supported on oslo.config.
Multiple APIs receives an image object now, which is an object
with the options provided in each image section.