Add support for inheriting scenarios in the scenario config.
Having this feature will allow to define succintly a new scenario,
without too much redundancy, which will result in simpler production
configurations when we'll have lots of scenarios for all supported
OSes.

To inherit a scenario, this readable syntax is used:

  [scenario inherits base scenario]

If any attribute is not found in a scenario, a lookup in the parent
is done, if any parent is defined. Otherwise an error will be raised
or the default value for that attribute will be returned.