Add support for grabbing the installation log.
Sometimes it is useful to introspect the installation log for various
errors and situations. This patch extends the Scenario class to accept
an additional parameter, output_directory, which will be a folder
where the installation log can be saved, as well as the instances
outputs, the same location being reused.