Fix instance output log retrieval
There's a minor issue with the `instance_output` method from
the API as it returns a number of output lines lower than
the requested one when the limit is reached. So the algorithm
is not requesting any more lines of output, because is considered
that the entire content was already retrieved (due to the higher limit).