Fix public keys EOL consistency
Every public keys set retrieved from the metadata should be a string
ended in and using EOL as a separator. Every metadata service in
cloudbaseinit will parse and obtain a list of stripped keys, which
will be written eventually in a file where each key is separated by EOL again.
This way, the standard is clear for both argus and cloudbaseinit + services:
    * every single public key is a simple string without
      leading or ending spaces
    * every string/content, with public keys, use EOL as separator
      and may end (metadata) with EOL (splitlines method will
                                       do the trick)