Send a proper certificate as MaaS metadata
The old x509 certificate was in fact a PEM with the public key,
which clearly was an invalid certificate, therefore the UPN related error.
Also, the certificate must be provided with a new line too (metadata standard)
and the x509 importer utility from cb-init will take care of encodings/EOLs.