Mark scenarios which failed with ArgusError as failed
This patch catches ArgusError when running a scenario and marks
its tests as failed. No other error is caught, since we believe
that other errors should be fixed ASAP, so stopping after the first
failing scenario is good for us, but ArgusError are usually under
constraints for one scenario or another. One use case for this is when
we're installing git and it fails due to network connectivity issues.