Add a script for stopping nicely the devstack environment
Sending the kill command to the devstack screen is not enough,
since some processes can linger on, blocking the use of the ports
they were using, which leads to scenarios being blocked, waiting
for the environment to be up again after a restart. This patch adds
a command `stop_devstack`, which is used instead of the screen
command.