Remove tests that were driven from the installation script
Some tests were relying on the fact that the script folder was
created by the installation script. Since it doesn't make sense to be
there, that particular code was removed and the additional tests were
removed as well. We did this since scripts are already tested using
the multipart userdata, so nothing is really lost.