Patch the installation of cloudbaseinit to create a file when finishing
We're doing this in order to verify that the service started and finished
executing, instead of just checking the service status. The service status
verification has the drawback that on some OSes and it some situations,
the service can start later on, leading argus to believe that it actually
finished, which means that the tests are executed too soon.