Raise error if cbinit dir is not found.
In some cases (for example when using the x64 version of cbinit on
a x86 os) that method will return None. That will cause ntpath.join
in set_config_option to fail with 'len(p) > 0 None type object has
no attribute len' which is not very informative so it's better if we
don't let the error pass unnoticed.