Add create_keypair method to APIManager class
The new method returns a new Keypair object, which wraps the
(public, private) keypair and provides a convenience method
for destroying the keypair. This change also brought a new
change in tempest_backend.py, where the key is now saved explicitly
into a new temporary file, instead of saving it all the time
in path_to_private_key.