Adds support for Windows 2008/2008R2
Adds support for versions of Windows 2008 and 2008R2.
A few things are being changed in the commands structure
so that we can have common and compatible commands across
all present supported versions of Windows in Argus.
Some of the notable changes:
* using the 'webclient' object for downloading resources in the underlying
instance, since it's also faster and more cross-compatible than
'Invoke-WebRequest'
* adds a sanitization method for the ouput returned by running commands
in the instance.
* the method for determining the os version is now using
direct values from the current environment, instead of calling
powershell commandlets.
* the MTU setter will now set the interfaces found on the underlying
instance, instead of always setting a constant name value which changes
across platforms.
* replaces the 'Ignore' argument in powershell scriptlets with
'SilentlyContinue', since 'Ignore' was introduced in version 3.0
of Powershell.
* adds a parser for ssh public keys, in order to make sure that
the content file got is correct.