Make ScenarioNetworkConfig use config drive
The following changes are required for the OpenStack backend because
the debian config does not contain information regarding the IPv6
configuration of the instance when the HTTP metadata service is used.
Since the IPv6 information is present only in the config drive, besides
the other networking details, the NetworkConfig scenario will be
configured to use the config drive instead of HTTP metadata service.