Renamed scenarios in order to control their independent execution
Tempest scenarios can be given as a regex list in this fashion:
'ScenarioSmoke|ScenarioIndependentPlugins'

In this case, all the scenarios that start with ScenarioSmoke will
be executed, which is not the wanted behaviour at times.
There is no way to just run only ScenarioSmoke if there is another
scenario like ScenarioSmokeGeneric (both will be executed).

This commit renames the scenarios that have names that fit the
above pattern, so that we can have granular control over each
scenario.